+++ Changelog.md
@@ -13,6 +13,7 @@
 
 ### Refactor
 * Move parsing Columns inside class
+* Move parsing TableGrid inside class
 
 ## 0.1.2 (2016-06-07)